pubsub_proto/requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
dill==0.2.7.1
future==0.16.0
gapic-google-cloud-pubsub-v1==0.15.4
gcloud==0.18.3
google-api-core==1.1.2
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-cloud-core==0.28.1
google-cloud-pubsub==0.34.0
google-gax==0.16.0
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.12.0
httplib2==0.11.3
idna==2.6
oauth2client==3.0.0
ply==3.8
proto-google-cloud-pubsub-v1==0.15.4
protobuf==3.5.2.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2018.4
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
dill==0.2.7.1
future==0.16.0
gapic-google-cloud-pubsub-v1==0.15.4
gcloud==0.18.3
google-api-core==1.1.2
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-cloud-core==0.28.1
google-cloud-pubsub==0.34.0
google-gax==0.16.0
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.12.0
httplib2==0.11.3
oauth2client==3.0.0
ply==3.8
proto-google-cloud-pubsub-v1==0.15.4
protobuf==3.5.2.post1
pyasn1==0.4.2
six==1.11.0
